Move source code to module_backend_t.c
