+++ example/CMakeLists.txt
@@ -3,9 +3,7 @@ set(ECOMMON_LIBRARIES common elliptics_client)
 
 set(DNET_IOSERV_SRCS ioserv.c config.c file_backend.c backends.c eblob_backend.c
  module_backend/core/module_backend_t.c
- module_backend/core/module_backend_config_t.c
  module_backend/core/dlopen_handle_t.c
- module_backend/core/module_backend_api_t.c
 )
 add_subdirectory(module_backend)
 set(DNET_IOSERV_LIBRARIES ${ECOMMON_LIBRARIES} elliptics elliptics_cocaine dl)
